AIchatbot-week3d3/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
together
langchain-google-genai
langchain-community
google-generativeai>=0.6.0
motor
pymongo
PyMuPDF
python-docx
pandas
openpyxl
python-multipart